--- a/.github/workflows/mypy.yml
+++ b/.github/workflows/mypy.yml
@@ -3,7 +3,7 @@ on:
   pull_request:
     branches:
       - master
-      - '**dev'
+      - "**dev"
     paths:
       - splink/**
       - tests/**
@@ -17,12 +17,13 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v4
 
-      - name: Setup uv
-        id: setup-uv
-        uses: astral-sh/setup-uv@v2
+      - name: Setup uv (with cache)
+        uses: astral-sh/setup-uv@v6
         with:
           enable-cache: true
-          cache-dependency-glob: "uv.lock"
+          cache-dependency-glob: |
+            pyproject.toml
+            uv.lock
 
       - name: Install Python 3.10
         run: uv python install 3.10