--- a/.github/workflows/pytest_spark.yml
+++ b/.github/workflows/pytest_spark.yml
@@ -24,10 +24,12 @@ jobs:
 
       - name: Setup uv
         id: setup-uv
-        uses: astral-sh/setup-uv@v2
+        uses: astral-sh/setup-uv@v6
         with:
           enable-cache: true
-          cache-dependency-glob: "uv.lock"
+              cache-dependency-glob: |
+      pyproject.toml
+      uv.lock
 
       - name: Install Python ${{ matrix.python-version }}
         run: uv python install ${{ matrix.python-version }}