--- a/.github/workflows/pytest_spark.yml
+++ b/.github/workflows/pytest_spark.yml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
         # pyspark < 4 does not support python 3.12
-        python-version: ["3.8", "3.9", "3.10"]
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
     name: Run tests with Python ${{ matrix.python-version }}
     steps:
       - uses: actions/checkout@v4