--- a/.github/workflows/run_demos_tutorials.yml
+++ b/.github/workflows/run_demos_tutorials.yml
@@ -26,12 +26,10 @@ jobs:
 
       - name: Setup uv
         id: setup-uv
-        uses: astral-sh/setup-uv@v6
+        uses: astral-sh/setup-uv@v2
         with:
           enable-cache: true
-              cache-dependency-glob: |
-      pyproject.toml
-      uv.lock
+          cache-dependency-glob: "uv.lock"
 
       - name: Install Python ${{ matrix.python-version }}
         run: uv python install ${{ matrix.python-version }}