… tests/test_debug_mode.py