Closes #2791 .  Following #2789 
Makes test suite run a lot faster.  

Before: 18 minutes
After: 6 minutes

On my mac, whole suite runs in  6 minutes.  Or < 2 minutes 36 seconds with `pytest-xdist` to run them in parallel

- Upgrade `astral-sh/setup-uv@v2` to `astral-sh/setup-uv@v6` so cache works again
- Reduce sample size on slow sqlite tests to make them run faster
- Stop running a couple of tests against Spark that weren't testing Spark-specific functionality and were slow 

We should probably remove the 3.8 tests, but that cain wait for another PR 

## Before:

### Sqlite:

```
============================== slowest durations ===============================
215.06s call     tests/test_cluster_at_multiple_thresholds.py::test_cluster_at_multiple_thresholds[sqlite-5000]
77.56s call     tests/test_cluster_using_single_best_links.py::test_single_best_links_one_to_one[sqlite]
58.06s call     tests/test_u_train.py::test_u_train_link_only_sample[sqlite]
24.98s call     tests/test_settings_options.py::test_model_heavily_customised_settings[sqlite]
22.77s call     tests/test_debug_mode.py::test_debug_mode_combined_training[sqlite]
19.03s call     tests/test_debug_mode.py::test_debug_mode_ptrrm_train[sqlite]
18.76s call     tests/test_debug_mode.py::test_debug_mode_u_training[sqlite]
10.89s call     tests/test_full_example_sqlite.py::test_full_example_sqlite
```

### Postgres

```
============================== slowest durations ===============================
22.71s call     tests/test_debug_mode.py::test_debug_mode_combined_training[postgres]
18.36s call     tests/test_debug_mode.py::test_debug_mode_u_training[postgres]
18.32s call     tests/test_debug_mode.py::test_debug_mode_ptrrm_train[postgres]
7.90s call     tests/test_full_example_postgres.py::test_full_example_postgres
6.37s call     tests/test_cluster_using_single_best_links.py::test_single_best_links_one_to_one[postgres]
5.94s call     tests/test_cluster_at_multiple_thresholds.py::test_cluster_at_multiple_thresholds[postgres-5000]
3.53s call     tests/test_find_new_matches.py::test_matches_work[postgres]
3.46s call     tests/test_settings_options.py::test_model_heavily_customised_settings[postgres]
```

### Duckdb

```
Coverage XML written to file coverage.xml
============================== slowest durations ===============================
9.28s call     tests/test_full_example_duckdb.py::test_full_example_duckdb
4.27s call     tests/test_debug_mode.py::test_debug_mode_combined_training[duckdb]
3.93s call     tests/test_full_example_duckdb.py::test_small_example_duckdb
3.05s call     tests/test_debug_mode.py::test_debug_mode_interactive_match_weights_chart[duckdb]
2.86s call     tests/test_debug_mode.py::test_debug_mode_profile_columns[duckdb]
2.61s call     tests/test_debug_mode.py::test_debug_mode_interactive_m_u_chart[duckdb]
2.20s call     tests/test_new_db_api.py::test_charts[duckdb]
2.06s call     tests/test_find_new_matches.py::test_matches_work[duckdb]
2.06s call     tests/test_u_train.py::test_u_train_link_only_sample[duckdb]
2.03s call     tests/test_profile_data.py::test_profile_using_duckdb
```


### Spark

```
============================== slowest durations ===============================
45.51s call     tests/test_debug_mode.py::test_debug_mode_combined_training[spark]
23.03s call     tests/test_cluster_at_multiple_thresholds.py::test_cluster_at_multiple_thresholds_mw_prob_equivalence[spark]
21.17s call     tests/test_debug_mode.py::test_debug_mode_ptrrm_train[spark]
20.80s call     tests/test_analyse_blocking.py::test_analyse_blocking_slow_methodology[spark]
19.45s call     tests/test_debug_mode.py::test_debug_mode_u_training[spark]
16.89s call     tests/test_debug_mode.py::test_debug_mode_profile_columns[spark]
15.38s call     tests/test_full_example_spark.py::test_full_example_spark
14.58s call     tests/test_debug_mode.py::test_debug_mode_em_training[spark]
14.41s call     tests/test_debug_mode.py::test_debug_mode_interactive_match_weights_chart[spark]
14.05s call     tests/test_debug_mode.py::test_debug_mode_interactive_m_u_chart[spark]
13.64s call     tests/test_new_db_api.py::test_full_run[spark]
13.08s call     tests/test_analyse_blocking.py::test_blocking_records_accuracy[spark]
13.02s call     tests/test_debug_mode.py::test_debug_mode_cluster_studio[spark]
12.05s call     tests/test_blocking.py::test_simple_end_to_end[spark]
11.68s call     tests/test_debug_mode.py::test_debug_mode_clustering[spark]
```

<img width="1029" height="1265" alt="image" src="https://github.com/user-attachments/assets/10ca07c3-b546-4ebd-850c-1f9942abf483" />