@RobinL 2025-10-03
### Duckdb
```
10.22s call     tests/test_full_example_duckdb.py::test_full_example_duckdb
4.10s call     tests/test_full_example_duckdb.py::test_small_example_duckdb
2.62s call     tests/test_u_train.py::test_u_train_link_only_sample[duckdb]
2.44s call     tests/test_new_db_api.py::test_charts[duckdb]
2.30s call     tests/test_find_new_matches.py::test_matches_work[duckdb]
2.09s call     tests/test_new_db_api.py::test_full_run[duckdb]
```