@RobinL 2025-10-03
Python 3.8 was end of life `2024-10-07`, 3.9 will be end of life end of October 2025