@RobinL 2025-10-03
It's building duckdb from source on 3.8 AND it's not using the cache